requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.8.0
django-extensions==3.1.3
django-filter==2.4.0
django-shell-plus==1.1.7
django-username-email==2.4.2
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
